+++ src/main/java/com/example/du_an_1/entity/MauSac.java
@@ -12,9 +12,9 @@ import lombok.*;
 @Entity
 @Table(name = "mausac")
 public class MauSac {
-@Id
+    @Id
  @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id ;
+    private Long id ;
     private String ma ;
     private String ten ;
