+++ src/main/java/com/example/du_an_1/repository/ChatLieuRepository.java
@@ -3,5 +3,5 @@ package com.example.du_an_1.repository;
 import com.example.du_an_1.entity.ChatLieu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChatLieuRepository extends JpaRepository<ChatLieu,Integer> {
+public interface ChatLieuRepository extends JpaRepository<ChatLieu,Long> {
 }
