+++ src/main/java/com/example/du_an_1/repository/DongSPRepository.java
@@ -3,5 +3,5 @@ package com.example.du_an_1.repository;
 import com.example.du_an_1.entity.DongSP;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DongSPRepository extends JpaRepository<DongSP,Integer> {
+public interface DongSPRepository extends JpaRepository<DongSP,Long> {
 }
