+++ src/main/java/com/example/du_an_1/repository/KichCoRepository.java
@@ -3,5 +3,5 @@ package com.example.du_an_1.repository;
 import com.example.du_an_1.entity.KichCo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface KichCoRepository extends JpaRepository<KichCo,Integer> {
+public interface KichCoRepository extends JpaRepository<KichCo,Long> {
 }
