+++ src/main/java/com/example/du_an_1/repository/NSXRepository.java
@@ -3,5 +3,5 @@ package com.example.du_an_1.repository;
 import com.example.du_an_1.entity.NSX;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NSXRepository  extends JpaRepository<NSX,Integer> {
+public interface NSXRepository  extends JpaRepository<NSX,Long> {
 }
