+++ src/main/java/com/example/du_an_1/repository/SanPhamRepository.java
@@ -3,5 +3,5 @@ package com.example.du_an_1.repository;
 import com.example.du_an_1.entity.SanPham;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SanPhamRepository extends JpaRepository<SanPham,Integer> {
+public interface SanPhamRepository extends JpaRepository<SanPham,Long> {
 }
