+++ src/main/java/com/example/du_an_1/repository/ThuongHieuRepository.java
@@ -3,5 +3,5 @@ package com.example.du_an_1.repository;
 import com.example.du_an_1.entity.ThuongHieu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ThuongHieuRepository extends JpaRepository<ThuongHieu,Integer> {
+public interface ThuongHieuRepository extends JpaRepository<ThuongHieu,Long> {
 }
